+++ src/kernel/mailbox.c
@@ -67,7 +67,7 @@ static unsigned int place(struct mailbox *mailbox, unsigned int event, unsigned 
             unsigned int slot = mailbox->head % MESSAGE_SLOTS;
             struct message *message = &mailbox->messages[slot];
 
-            message_init(message, event, source, count, (void *)(KERNEL_VMAILBOX + MESSAGE_SIZE * slot + MESSAGE_CAPACITY * mailbox->ichannel));
+            message_init(message, event, source, count, (void *)(unsigned long)(KERNEL_VMAILBOX + MESSAGE_SIZE * slot + MESSAGE_CAPACITY * mailbox->ichannel));
             buffer_copy((void *)(mailbox->data + MESSAGE_SIZE * slot), data, count);
 
             mailbox->head++;
